fixed ambigous overloads is_power_of_2/next_pow2 for uint32/uint64 vs size_t

+++ crnlib/crn_vector.cpp
@@ -21,13 +21,13 @@ namespace crnlib
       if (m_capacity >= min_new_capacity)
          return true;
 
-      size_t new_capacity = min_new_capacity;
+      uint64 new_capacity = uint64(min_new_capacity);
       if ((grow_hint) && (!math::is_power_of_2(new_capacity)))
          new_capacity = math::next_pow2(new_capacity);
 
       CRNLIB_ASSERT(new_capacity && (new_capacity > m_capacity));
 
-      const size_t desired_size = element_size * new_capacity;
+      const size_t desired_size = element_size * size_t(new_capacity);
       size_t actual_size;
       if (!pMover)
       {
